+++ nrfutil/index.ts
@@ -4,8 +4,7 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-export { default as prepareSandbox } from './sandbox';
-export type { NrfutilSandbox } from './sandbox';
+export { NrfutilSandbox } from './sandbox';
 export type { Progress } from './sandboxTypes';
 export { getNrfutilLogger, setNrfutilLogger } from './nrfutilLogger';
 export {